pyproject.toml
[project]
name = "medigraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "bitsandbytes>=0.47.0",
    "fastapi>=0.135.1",
    "peft>=0.17.1",
    "pillow>=11.3.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "tqdm>=4.67.1",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
    "pandas>=3.0.2",
    "evaluate>=0.4.6",
    "jiwer>=4.0.0",
    "datasets>=4.8.4",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "neo4j>=6.1.0",
    "langchain>=1.2.15",
    "langchain-core>=1.2.28",
    "langchain-neo4j>=0.9.0",
    "langchain-ollama>=1.1.0",
    "ollama>=0.6.0",
    "python-dotenv>=1.2.2",
    "biopython>=1.87",
    "mygene>=3.2.2",
]
